+++ src/logan/pickpocket/commands/ProfileRemoveCommand.kt
@@ -1,0 +1,22 @@
+package logan.pickpocket.commands
+
+import logan.api.command.BasicCommand
+import logan.api.command.SenderTarget
+import org.bukkit.entity.Player
+
+class ProfileRemoveCommand : BasicCommand<Player>(
+    "remove",
+    2..2,
+    listOf(String::class, String::class),
+    "profile",
+    SenderTarget.PLAYER,
+    "pickpocket.admin.profile.remove",
+    """
+        Usage:
+        /pickpocket profile remove <thief|victim> <name>
+    """.trimIndent()
+) {
+    override fun run(sender: Player, args: Array<out String>, data: Any?): Boolean {
+        TODO()
+    }
+}
