+++ src/logan/pickpocket/main/ProfileType.kt
@@ -1,0 +1,7 @@
+package logan.pickpocket.main
+
+enum class ProfileType {
+    THIEF, VICTIM;
+
+    val friendlyName = name.lowercase()
+}
