+++ src/logan/pickpocket/main/VictimProfile.kt
@@ -1,0 +1,6 @@
+package logan.pickpocket.main
+
+class VictimProfile(override val name: String) : Profile {
+    override val type = ProfileType.VICTIM
+    override val properties = mutableMapOf<String, Any>()
+}
